requirements.txt
Django==1.6.9
GitPython==0.3.2.RC1
Markdown==2.3.1
Pillow==2.7.0
South==1.0.2
astroid==1.2.0
async==0.6.1
boto==2.24.0
cov-core==1.14.0
coverage==3.7.1
dj-database-url==0.2.2
dj-static==0.0.5
django-admin-tools==0.5.2
django-dbbackup==1.80.2
django-debug-toolbar==1.0.1
django-filter==0.7
django-guardian==1.1.1
django-heroku-memcacheify==0.5
django-mathfilters==0.2.1
django-nose==1.2
django-pylibmc-sasl==0.2.4
django-toolbelt==0.0.1
django-userena==1.4.0
djangorestframework==3.0.5
docopt==0.4.0
docutils==0.11
drf-nested-routers==0.1.3
dropbox==2.0.0
easy-thumbnails==1.3
gdata==2.0.18
gitdb==0.5.4
google-api-python-client==1.2
gunicorn==18.0
heroku==0.1.4
html2text==3.200.3
httplib2==0.8
humanize==0.5
logilab-common==0.62.1
mailchimp==2.0.8
mandrill==1.0.53
mock==1.0.1
newrelic==2.12.0.10
nose==1.3.4
nose-cov==1.6
psycopg2==2.5.1
pylibmc==1.2.3
pylint==1.3.0
python-dateutil==1.5
python-gflags==2.0
pytz==2013.9
redis==2.9.1
requests==2.2.1
rq==0.3.13
simplejson==3.5.2
six==1.5.2
smmap==0.8.2
sqlparse==0.1.10
static==0.4
stripe==1.14.0
times==0.6.2
urllib3==1.8.2
validate-email==1.2
wsgiref==0.1.2
yolk==0.4.3
